Neural-Nodes/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
tensorflow==2.15.0
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.1
scikit-learn==1.4.0